requirements.txt
django==3.2.5
djangorestframework==3.12.4
django-widget-tweaks==1.4.8
psycopg2-binary==2.9.1
